+++ src/view/index.ts
@@ -1,0 +1,2 @@
+import './ActorAnimations'
+import './ItemAnimations'